Harden image/library/MCP paths against malformed input and overwrites
Eight runtime defects found in a full-project sweep, all in pure
(non-Qt) logic and covered by new unit tests:

- Dated-folder import checked only in-batch name collisions, so a photo
  whose name already existed in the target folder was silently
  overwritten. Planning now avoids on-disk collisions too, and a failed
  copy/move skips that file instead of aborting the whole import.
- XMP sidecar load caught only ParseError/OSError; a sidecar carrying a
  DTD or entity declaration raised defusedxml's DefusedXmlException (a
  ValueError subclass), crashing keyword indexing and smart-album
  evaluation. It now degrades to an empty sidecar.
- .cube LUT header parsing raised IndexError on a directive with no
  value (truncated file), escaping callers' (OSError, ValueError) guard
  and crashing the develop render. Missing size / domain tokens now
  raise ValueError per the documented contract.
- Error-level-analysis amplified the JPEG diff in int16, which overflows
  for scale >= 129 and wrapped the strongest tamper signal to black.
  Amplification is now done in int32.
- Local-adjustment brush masks assumed every point was a dict; a
  list-shaped point in a persisted recipe raised AttributeError past the
  recipe guard. Malformed points are skipped.
- XMP export dropped dc:creator (Imervue doesn't track it), wiping a
  creator written by an external editor and deleting creator-only
  sidecars. The value is now carried over from the existing sidecar.
- convert_format failed on palette-mode ("P") sources for JPEG/BMP
  ("cannot write mode P as JPEG"); non-RGB/L modes are now flattened.
- resize_image and build_collage clamped dimensions only at the low end,
  so an unvalidated request could allocate ~10 GB and block the
  synchronous MCP server. Output is now capped at 100 MP.

Adds 21 regression tests.

+++ Imervue/image/ela.py
@@ -43,7 +43,10 @@ def error_level_analysis(
         baseline = np.asarray(recompressed.convert("RGB"), dtype=np.int16)
 
     diff = np.abs(rgb.astype(np.int16) - baseline)
-    amplified = np.clip(diff * scale, 0, 255).astype(np.uint8)
+    # Amplify in int32: diff (0..255) * scale overflows int16 once scale >= 129
+    # (255*129 > 32767), wrapping the strongest tamper signal to a negative
+    # value that then clips to black — the exact opposite of the intent.
+    amplified = np.clip(diff.astype(np.int32) * scale, 0, 255).astype(np.uint8)
     out = np.empty((*rgb.shape[:2], _RGBA_CHANNELS), dtype=np.uint8)
     out[..., :3] = amplified
     out[..., 3] = _OPAQUE
